refactor(ReadComicOnline): update to reflect site branding
