+++ src/ReadComicOnline/implementations/discover-section/parsers.ts
