+++ src/ReadComicOnline/implementations/manga/parsers.ts
