+++ src/ReadComicOnline/implementations/search-results/parsers.ts
