+++ src/ReadComicOnline/implementations/settings-form/forms/landing.ts
@@ -2,7 +2,7 @@ import { Form, NavigationRow, Section, type FormSectionElement } from "@paperbac
 import { DiscoverSettingsForm } from "./discover";
 import { SearchSettingsForm } from "./search";
 
-export class ReadComicOnlineLiSettingsForm extends Form {
+export class ReadComicOnlineSettingsForm extends Form {
   override getSections(): FormSectionElement<unknown>[] {
     return [
       Section("mainSettings", [
