+++ src/ReadComicOnline/implementations/shared/models.ts
